CGLWindow review: Now it's possible to change the resolution and fullscreenness on the fly (Fixes issue #44). Just set needed parameters and call CGLWindow::Initialize, that will do the trick for you. Resizing and maximizing a window also work now. Added few getters, such as GetFullscreen, GetBPP, etc.
Added Reload function to CResource, that is used to unload and then load again any currently loaded resource.
Added ReloadTextures function to CTextureManager.
Changed the logic of attributes checks in deserialization: now they do nothing if an attribute value is in an improper format.
Added Detach function to the Lua API.